完赛

+++ Camellia2/Camellia/USER/Hardware/Motor.h
@@ -53,8 +53,8 @@ extern int RMotor_PWM;
 
 void Motor_Init(void);
 void Motor_PWM(int L_PWM,int R_PWM);
-int16 limit(int16 In,int16 limit);
-void Motor_SET_PID(float Kp,float Ki,float Kd);
+int32 limit(int32 In, int32 limit);
+void Motor_SET_PID(float Kp, float Ki, float Kd);
 int Speed_pid_Out(int Target_Value,int Actual_Value);
 void Rotary_Plug(T);
